added validations for writer acces logic

+++ controller/ControllerWriter.php
@@ -17,11 +17,14 @@ class ControllerWriter extends Controller{
     public function create(){
         //this should not be available if you 
         //are logged in already
-        if($_SESSION['fingerPrint']){
+        if(isSet($_SESSION['fingerPrint'])){
             Twig::render('home-error.php', ['message'=> "You already seem to have an account ;P"]);
             return;
         }
-        Twig::render('writer-create.php');
+
+        $privilege = new Privilege;
+        $select = $privilege->select();
+        Twig::render('writer-create.php', ['privilege' => $select]);
     }
 
     public function store(){
@@ -33,13 +36,14 @@ class ControllerWriter extends Controller{
 
         //this should not be available if you 
         //are logged in already
-        if($_SESSION['fingerPrint']){
+        if(isSet($_SESSION['fingerPrint'])){
             Twig::render('home-error.php', ['message'=> "You already have an account ;P"]);
             return;
         }
 
         $this->validateWriter($_POST);
         extract($_POST);
+        unset($_POST['currentPage']);
 
         $writer = new Writer;
         $options =[
@@ -49,19 +53,14 @@ class ControllerWriter extends Controller{
         $passwordHash = password_hash($password, PASSWORD_BCRYPT, $options);
         $_POST['password']= $passwordHash;
         $insert = $writer->insert($_POST);
-        //Twig::render('writer-index.php');
         RequirePage::redirect('writer');
     }
 
     public function show($id = null){
-        //TO-DO: you want all the information available
-        //to admins. writers should only have access to bios...
-        //and eventually to associated texts. 
-
         $writer = new Writer;
         $selectId = $writer->selectId($id);
 
-        //here, you check if the id doesn't exist, or if it's null
+        //error page if id doesn't exist, or if it's null
         if($id == null || !$selectId){
             Twig::render('home-error.php', ['message'=> "Something went wrong. Sorry."]);
         }else{
@@ -69,21 +68,42 @@ class ControllerWriter extends Controller{
         }
     }
 
-    public function edit($id){
-        //TO-DO: you want this only available to admins
-        //and to the person who's id matches the writer-id
-        //I don't think I should be sending the id to be edited
-        //as a get...
+    public function edit(){
+        //check if the writer is logged in
+        CheckSession::sessionAuth();
 
+        //make sure the page is being accessed with a post...
+        if($_SERVER["REQUEST_METHOD"] !== "POST"){
+            //if logged in, direct to texts..
+            RequirePage::redirect('writer');
+            exit();
+        }
+        $privilege = new Privilege;
+        $select = $privilege->select();
         $writer = new Writer;
-        $selectId = $writer->selectId($id);
-        Twig::render('writer-edit.php', ['writer' => $selectId]);
+        $selectId = $writer->selectId($_POST['id']);
+        Twig::render('writer-edit.php', ['data' => $selectId, 'privilege' => $select]);
     }
 
     public function update(){
-        //TO-DO: you want this only available to admins
+        //only available to admins 1
         //and to the person who's id matches the writer-id
+        if($_SESSION['writer_id'] != $_POST['id'] && $_SESSION['privilege'] != 1 ){
+            Twig::render('home-index.php');
+            exit();
+        }
 
+        //make sure the page is being accessed with a post...
+        if($_SERVER["REQUEST_METHOD"] !== "POST"){
+            RequirePage::redirect('writer');
+            exit();
+        }
+
+        //validate
+        $this->validateWriter($_POST);
+
+        //clean the $_POST for the update
+        unset($_POST['currentPage']);
         $writer = new Writer;
         $update = $writer->update($_POST);
         $id = $_POST['id'];
@@ -91,34 +111,55 @@ class ControllerWriter extends Controller{
     }
 
     public function destroy(){
-        //TO-DO: you want this only available to admins
+        //only available to admins 1
         //and to the person who's id matches the writer-id
+        if($_SESSION['writer_id']  !== $_POST['id'] && $_SESSION['privilege'] !== 1 ){
+            RequirePage::redirect('writer');
+            exit();
+        }
+
+        //make sure the page is being accessed with a post...
+        if($_SERVER["REQUEST_METHOD"] !== "POST"){
+            RequirePage::redirect('writer');
+            exit();
+        }
         
         $writer = new Writer;
         $delete = $writer->delete($_POST['id']);
+        $id = $_SESSION['writer_id'];
         if($delete){
+            if($_POST['id'] == $id){
+                session_destroy();
+                RequirePage::redirect('login');
+                exit();
+            }
             RequirePage::redirect('writer');
         }else{
             Twig::render('home-error.php', ['message'=> "To delete this writer, you must first delete all their texts--which will only be possible if no one has iterated on them. We would recommend you deactivate this account instead, but we have not yet implemented this functionality. Sorry."]);
         }
     }
 
-    public function validateWriter($data){
+    public function validateWriter($data, $password="default"){
         extract($data);
+
         RequirePage::library('Validation');
         $val = new Validation;
         $val->name('firstName')->value($firstName)->required()->max(45)->pattern('words');
         $val->name('lastName')->value($lastName)->required()->max(45)->pattern('words');
-        $val->name('username')->value($email)->pattern('email')->required()->max(50);
+        $val->name('email')->value($email)->pattern('email')->required()->max(50);
         $val->name('password')->value($password)->pattern('alphanum')->required()->min(6)->max(20);
+        //TODO: put a db default value so that you don't need to put it required
+        $val->name('birthday')->value($birthday)->required();
 
         if($val->isSuccess()){
             //continue
         }else{
+
             $errors = $val->displayErrors();
             $privilege = new Privilege;
-            $select = $privilege->select();
-            Twig::render('writer-create.php', ['privileges' => $select, 'errors' => $errors, 'data' => $_POST]);
+            $privileges = $privilege->select();
+            Twig::render($currentPage, ['privilege' => $privileges, 'errors' => $errors, 'data' => $_POST]);
+            exit();
         };
     }
